First and Final CSS commit
Changed mw_styles_txt.css to my_styles.css.  Added the links for the mw_layout.css and the mw_styles.css. The style rule is for the html and the background color is rgb(173, 189, 227). The style rule is for the body element and it says that the background color is rgb(227, 210, 173) and the font family is century Gothic, sans-serif. This rule is for the body footer and the background color is rgb(208, 184, 109) and the padding is for the top and the bottom and its 5 pixels. This rule is for the address element and the font style is normal, font size is 0.9 pixels, text-align is horizontal center, the padding is for the top/bottom 10 pixels. This rule is for the primary headings and the font weight is normal.  This rule is for the secondary headings and the font weight is normal. This rule is for the Nav and has the unordered list nested in the nav and it states the list style type is none, the line height is 2em and the font size is 0.9em.  This rule is for the nav and the unordered list that is nested in the nav and it for the links that are being visited and the color is rgb(43, 59, 125) and there are no text decoration.  This rule is for the nav and the unordered list thats nested in the nav and the it is saying that for all of the links that are being hover by will changed to the color red. The rule is saying that the header is in the id and the primary heading is in the header and the font size os 2.2 em and the letter spacing is 0.2em. The rule is saying that the header is in the id and the paragraphs are in the header and the font size is 1.1em. The rule is stating that the paragraphs are directly in the id and that the first paragraph and the first line will have small caps and the font size will be 1.2em.  The primary heading is in the id and the font size will be 1.2 em. The secondary heading is in the id and the font size will be 0.9em. The paragraphs are in the id and the font size is 0.8em. The secondary heading and the paragraphs and the odds are in the id and the color will be rgb(79, 91, 40).  The block-quote element has the background color set as rgb(173, 189, 227) and the color is rgba(255, 255, 255, 0.65).